mestrado/outros/PLDCA-master/pyproject.toml
[tool.poetry]
name = "pldca"
version = "0.1.0"
description = ""
authors = ["Gabriel Soares Baptista <gsoaresbaptista@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.4.0"
pyyaml = "^6.0.2"
isort = "^5.13.2"
torchvision = "^0.19.0"
black = "^24.8.0"
matplotlib = "^3.9.2"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
